refactor: `WalletConnect` -> `WalletConnectColor`
